requirements.txt
# ==================== Web 框架 ====================
fastapi==0.115.0
uvicorn[standard]==0.30.6

# ==================== 数据验证 ====================
pydantic==2.9.2
pydantic-settings==2.6.1

# ==================== 认证与安全 ====================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ==================== 数据库 ====================
sqlalchemy[asyncio]==2.0.23  # SQLAlchemy ORM（异步支持）
asyncpg==0.29.0  # PostgreSQL 异步驱动
alembic==1.13.1  # 数据库迁移工具

# ==================== 缓存 ====================
redis==5.2.1

# ==================== HTTP 客户端 ====================
aiohttp==3.11.11
httpx==0.28.1

# ==================== 文件处理 ====================
aiofiles==24.1.0
python-multipart==0.0.20

# ==================== 配置与日志 ====================
python-dotenv==1.0.1
pyyaml==6.0.2
loguru==0.7.3

# ==================== XML 解析 ====================
lxml==5.3.0

# ==================== 任务队列（已有 arq 配置）====================
arq==0.26.1  # 如果使用 arq 任务队列

# ==================== 工具库 ====================
pinyin==0.4.0  # 中文转拼音（用于生成 ID）
python-ulid==2.7.0  # ULID 生成器